moved logger and sanity_checker to /utils
