+++ lib/stimulus_reflex/utils/logger.rb
