+++ lib/stimulus_reflex/utils/sanity_checker.rb
